move files in a stella folder
