+++ lib/zakuro/version/senmyou/stella/lunar/location.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../const/const'
+require_relative '../../const/const'
 
 # :nodoc:
 module Zakuro
