+++ lib/zakuro/version/senmyou/stella/solar/average.rb
