+++ lib/zakuro/version/senmyou/stella/solar/location.rb
